Completed all functions

+++ src/java/utils/DataAccessObject.java
@@ -6,13 +6,15 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Timestamp;
 import java.util.ArrayList;
-import java.util.Collections;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 import model.Answer;
 import model.User;
 import model.Bank;
 import model.Question;
 import model.Exam;
+import model.Record;
 import model.Role;
 import org.apache.catalina.tribes.util.Arrays;
 
@@ -188,13 +190,14 @@ public class DataAccessObject {
         }
         return false;
     }
-    
-     public static boolean removeBank(int bankID) {
+
+    public static boolean removeBank(int bankID, int creatorID) {
         try {
             try (Connection conn = DBContext.getConnection()) {
-                String query = "delete from bankTbl where bankID=?";
+                String query = "delete from bankTbl where bankID = ? and creatorID = ?";
                 try (PreparedStatement stm = conn.prepareStatement(query)) {
                     stm.setInt(1, bankID);
+                    stm.setInt(2, creatorID);
                     return stm.executeUpdate() > 0;
                 }
             }
@@ -204,12 +207,16 @@ public class DataAccessObject {
         return false;
     }
 
-    public static boolean removeExam(String examCode) {
+    public static boolean removeExam(String examCode, int creatorID) {
         try {
             try (Connection conn = DBContext.getConnection()) {
-                String query = "delete from examTbl where examCode=?";
+                String query = "delete from examTbl where examCode = ? and creatorID = ?\n"
+                        + "delete from questionTbl\n"
+                        + "where bankID is null and questionID not in \n"
+                        + "(select questionID from examDetailTbl)";
                 try (PreparedStatement stm = conn.prepareStatement(query)) {
                     stm.setString(1, examCode);
+                    stm.setInt(2, creatorID);
                     return stm.executeUpdate() > 0;
                 }
             }
@@ -219,7 +226,55 @@ public class DataAccessObject {
         return false;
     }
 
-    public static List<Exam> getExam(String examCode) {
+    public static List<Record> getRecordsOfAExam(String examCode) {
+        List<Record> records = new ArrayList<>();
+        try {
+            try (Connection conn = DBContext.getConnection()) {
+                String query = "select rt.recordID, rt.takerID, ut.fullname, rt.examDate, rt.dateSubmitted, count(rdt.answerID) as numOfCorrectAnswers, count(at2.answerID) as numOfAnswers from recordTbl rt \n"
+                        + "inner join userTbl ut on rt.takerID = ut.userID \n"
+                        + "inner join examDetailTbl edt on rt.examCode = edt.examCode \n"
+                        + "inner join answerTbl at2 on edt.questionID = at2.questionID \n"
+                        + "left outer join recordDetailTbl rdt on at2.answerID = rdt.answerID\n"
+                        + "where at2.isCorrect = 1 and rt.examCode = ?\n"
+                        + "group by rt.recordID, rt.takerID, ut.fullname, rt.examDate, rt.dateSubmitted";
+
+                try (PreparedStatement stm = conn.prepareStatement(query)) {
+                    stm.setString(1, examCode);
+                    ResultSet resset = stm.executeQuery();
+                    while (resset.next()) {
+                        records.add(new Record(examCode, resset.getInt("recordID"), resset.getInt("takerID"), resset.getString("fullname"),
+                                resset.getTimestamp("examDate"), resset.getTimestamp("dateSubmitted"),
+                                resset.getInt("numOfCorrectAnswers"), resset.getInt("numOfAnswers")));
+                    }
+                }
+            }
+        } catch (SQLException ex) {
+            System.out.println(ex.getMessage() + "\n" + ex.getStackTrace());
+        }
+
+        return records;
+    }
+
+    public static boolean removeRecord(int recordID, int creatorID) {
+        try {
+            try (Connection conn = DBContext.getConnection()) {
+                String query = "delete from recordTbl \n"
+                        + "where recordID = ? and recordID in (select recordID from examTbl et \n"
+                        + "inner join recordTbl rt on et.examCode = rt.examCode \n"
+                        + "where et.creatorID = ?)";
+                try (PreparedStatement stm = conn.prepareStatement(query)) {
+                    stm.setInt(1, recordID);
+                    stm.setInt(2, creatorID);
+                    return stm.executeUpdate() > 0;
+                }
+            }
+        } catch (SQLException ex) {
+            System.out.println(ex.getMessage());
+        }
+        return false;
+    }
+
+    public static List<Exam> getExam(String examCode, int creatorID) {
         List<Exam> exams = new ArrayList<>();
         try {
             try (Connection conn = DBContext.getConnection()) {
@@ -227,11 +282,23 @@ public class DataAccessObject {
                         + "inner join examDetailTbl edt on et.examCode = edt.examCode \n"
                         + "inner join questionTbl qt on qt.questionID = edt.questionID \n"
                         + "inner join answerTbl at2 on qt.questionID = at2.questionID\n"
-                        + (examCode != null ? "where et.examCode = ?\n" : "")
-                        + "order by et.examCode, qt.questionID";
+                        + "inner join userTbl ut on et.creatorID = ut.userID\n";
+                if (examCode != null && creatorID > 0) {
+                    query += "where ut.userID = ? and et.examCode = ?\n";
+                } else if (examCode != null) {
+                    query += "where et.examCode = ?\n";
+                } else if (creatorID > 0) {
+                    query += "where ut.userID = ?\n";
+                }
+                query += "order by et.examCode, qt.questionID\n";
                 try (PreparedStatement stm = conn.prepareStatement(query)) {
-                    if (examCode != null) {
+                    if (examCode != null && creatorID > 0) {
+                        stm.setInt(1, creatorID);
+                        stm.setString(2, examCode);
+                    } else if (examCode != null) {
                         stm.setString(1, examCode);
+                    } else if (creatorID > 0) {
+                        stm.setInt(1, creatorID);
                     }
                     ResultSet resset = stm.executeQuery();
                     Exam exam = null;
@@ -246,7 +313,7 @@ public class DataAccessObject {
                             if (exam != null) {
                                 exams.add(exam);
                             }
-                            exam = new Exam(examCode, new ArrayList<>(), resset.getTimestamp("openDate"), resset.getTimestamp("closeDate"), null, resset.getInt("duration"));
+                            exam = new Exam(examCode, new ArrayList<>(), resset.getTimestamp("openDate"), resset.getTimestamp("closeDate"), resset.getString("fullName"), null, resset.getInt("duration"));
                             question = new Question(questionID, resset.getString("questionContent"), 0, new ArrayList<>());
                             answer = new Answer(answerID, resset.getString("answerContent"));
                             maxChoose = 0;
@@ -277,6 +344,79 @@ public class DataAccessObject {
         return exams;
     }
 
+    public static Set<Integer> getSelectedAnswers(int recordID, int creatorID) {
+        Set<Integer> selected = new HashSet<>();
+        try {
+            try (Connection conn = DBContext.getConnection()) {
+                String query = "select * from recordTbl rt \n"
+                        + "inner join recordDetailTbl rdt on rt.recordID = rdt.recordID\n"
+                        + "inner join examTbl et on rt.examCode = et.examCode \n"
+                        + "where rt.recordID = ? and et.creatorID = ?";
+                
+                try (PreparedStatement stm = conn.prepareStatement(query)) {
+                    stm.setInt(1, recordID);
+                    stm.setInt(2, creatorID);
+                    ResultSet resset = stm.executeQuery();
+                    while (resset.next()) {
+                        selected.add(resset.getInt("answerID"));
+                    }
+                }
+            }
+        } catch (SQLException ex) {
+            System.out.println(ex.getMessage());
+            ex.printStackTrace();
+        }
+        return selected;
+    }
+
+    public static List<Question> getQuestionsAndCorrectAnswer(String examCode, int creatorID) {
+        List<Question> questions = new ArrayList<>();
+        try {
+            try (Connection conn = DBContext.getConnection()) {
+                String query = "select * from examTbl et \n"
+                        + "inner join examDetailTbl edt on et.examCode = edt.examCode \n"
+                        + "inner join questionTbl qt on qt.questionID = edt.questionID \n"
+                        + "inner join answerTbl at2 on qt.questionID = at2.questionID\n"
+                        + "inner join userTbl ut on et.creatorID = ut.userID\n"
+                        + "where ut.userID = ? and et.examCode = ?\n"
+                        + "order by et.examCode, qt.questionID\n";
+                try (PreparedStatement stm = conn.prepareStatement(query)) {
+                    stm.setInt(1, creatorID);
+                    stm.setString(2, examCode);
+                    ResultSet resset = stm.executeQuery();
+                    Question question = null;
+                    Answer answer = null;
+                    int maxChoose = 0;
+                    while (resset.next()) {
+                        int questionID = resset.getInt("questionID");
+                        int answerID = resset.getInt("answerID");
+                        if (question == null || question.getQuestionID() != questionID) {
+                            if (question != null) {
+                                question.setMaxChoose(maxChoose);
+                                questions.add(question);
+                            }
+                            maxChoose = 0;
+                            question = new Question(questionID, resset.getString("questionContent"), 0, new ArrayList<>());
+                        }
+                        answer = new Answer(answerID, resset.getString("answerContent"));
+                        answer.setSelected(resset.getBoolean("isCorrect"));
+
+                        if (answer.isSelected()) {
+                            maxChoose++;
+                        }
+                        question.getAnswers().add(answer);
+                    }
+                    question.setMaxChoose(maxChoose);
+                    questions.add(question);
+                }
+            }
+        } catch (SQLException ex) {
+            System.out.println(ex.getMessage());
+            ex.printStackTrace();
+        }
+        return questions;
+    }
+
     public static List<Bank> getBank(int bankID) {
         List<Bank> banks = new ArrayList<>();
         try {
@@ -304,7 +444,7 @@ public class DataAccessObject {
                             if (bank != null) {
                                 banks.add(bank);
                             }
-                            bank = new Bank(bankID, resset.getString("bankName"), resset.getString("courseName"), resset.getString("fullName"), new ArrayList<>(), resset.getTimestamp("dateCreated"));
+                            bank = new Bank(bankID, resset.getString("bankName"), resset.getString("courseName"), resset.getInt("creatorID"), resset.getString("fullName"), new ArrayList<>(), resset.getTimestamp("dateCreated"));
                             question = new Question(questionID, resset.getString("questionContent"), 0, new ArrayList<>());
                             answer = new Answer(answerID, resset.getString("answerContent"));
                             maxChoose = 0;
@@ -330,7 +470,8 @@ public class DataAccessObject {
                 }
             }
         } catch (SQLException ex) {
-            System.out.println(ex.getMessage() + "\n" + ex.getStackTrace());
+            System.out.println(ex.getMessage());
+            ex.printStackTrace();
         }
         return banks;
     }
@@ -356,6 +497,7 @@ public class DataAccessObject {
             }
         } catch (SQLException ex) {
             System.out.println(ex.getMessage());
+            ex.printStackTrace();
         }
         return false;
     }
@@ -384,23 +526,9 @@ public class DataAccessObject {
             }
         } catch (SQLException ex) {
             System.out.println(ex.getMessage());
+            ex.printStackTrace();
         }
         return false;
-    }
-
-    public static void cleanDatabase() {
-        try {
-            try (Connection conn = DBContext.getConnection()) {
-                String query = "delete from questionTbl\n"
-                        + "where bankID is null and questionID not in \n"
-                        + "(select questionID from examDetailTbl)";
-                try (PreparedStatement stm = conn.prepareStatement(query)) {
-                    stm.executeUpdate();
-                }
-            }
-        } catch (SQLException ex) {
-            System.out.println(ex.getMessage());
-        }
     }
 
     public static boolean saveRecord(String examCode, int takerID, Exam exam, int selectedCount) {
